CareerAI/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
scikit-learn==1.3.2
xgboost==2.0.3
pandas==2.1.4
numpy==1.24.3
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
torch==2.1.1
transformers==4.36.0
plotly==5.17.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
